[refractor] post menu
